requirements.txt.txt
streamlit
tensorflow
opencv-python-headless
pillow
numpy
matplotlib
google-generativeai